fix: esm package traversal
when attempting to load the copilot nodejs will attempt to traverse
upwards to find a package.json to determin if it should run in ESM or
CJS mode.

if it finds one that contains `"type": "module"` it will crash copilot
because ESM handles various things different and the copilot impl is
CJS.

the solution here is to place a package.json file to prevent node from
traversing upwards.